merge facts and rules from all the blocks before validation
token verification was a bit conservative: it was evaluating each block
separately, restarting every time from a snapshot of the authority block.
We can instead merge all blocks before validation:
- caveats that apply on authority and ambient data should be independent
- authority and ambient facts and rules still cannot be provided by blocks
- rules like the revocation check will not need to differentiate between
the authority block and other blocks
- queries will now apply to the entire system instead of returning one
result per block validation
- the verifier could even provide the generated world for queries and
inspection after having validated it

The main issue I see right now is when a block relies on an intermediate
rule in a caveat. A later block could provide new facts that would be
used in that caveat. Something like this:
- block1 has special_right("X", #read) <- identity(admin) & resource(#ambient, "X")
- block1 has a caveat checking for special_right
- block2 provides the fact special°right("X", #read)

For this case, keeping the authority facts generation separated is
crucial