support caveats with multiple queries
a query can express AND conditions, but not OR. Some caveats might need
OR conditions (example: checking the presence of any of 2 different
facts with different format).

To implement it, we need caveats to hold multiple queries. If at east
one of those queries succeeds, the caveat succeeds.

This could be implemented with intermediate rules, but we would need to
check that blocks do not interfer with the results of those rules (by
pregenerating facts, etc)