add a way to create a verifier without a token
this will cover two use cases:
- validating policies when there's no token available
(unauthenticated user, etc)
- when a verifier needs a lot of facts, it can be preloaded with them,
then cloned for every requests, avoiding the work of converting them to
a datalog world generated from the token

There's an issue with incompatible symbol tables: if we pregenerate a
world for the verifier, it will have a different symbol table from the
token's, possibly sharing some symbols with different indexes.
Right now it is solved in a crude way: facts, rules and caveats coming
from the token are converted to their builder version with the token's
symbol table, then converted back to the datalog version using the
verifier's table.
There is a faster way, by comparing the to symbol tables and updating
the indexes in the terms. It will be implemented if there's too much
overhead (right now, the crypto part is much slower than everything
else)