implement allow/deny policies
caveats can be provided by the verifier or the token, and all of them
must be validated, while allow/deny policies are only provided by the
verifier, and we test them one by one until one matches.

This requires that the verifier always provides a default
rule,to allow or deny by default.