add a parser for source files
until now, the parser has been used mainly to transform a string passed as
argument to methods for token creation or the verifier, so we only
needed to call one parser for a fact, or one or a check, etc.

Now that the syntax is more usable, it makes more sense to represent a
list of facts, rules, checks and policies as one text file, with
comments, etc.

TBD:
- comments syntax
- span information for checks and policies (to highlight which ones were
triggered)