Fix date parsing in block contexts where items are semicolon-separated
The date parser first detects the end of the date token, before parsing it.
To that effect, it breaks on delimiter chars that cannot be contained in the
date literal. When using `parse_source` or `parse_block_source`, `;` is a possible
delimiter.

A more robust solution might be to use a positive condition (ie list all the chars that are
allowed in a date literal) instead.