implement scoped rules execution
in this model, rules, checks and policies only have access to facts
added or generated in the current or previous blocks, not later blocks.
Verifier rules, checks and policies are executed in the context of the
authority block.
With this change, we can get rid of #authority and #ambient, and
privileged rules do not need to exist