remove the symbol type, use the symbol table for strings
the symbol type was confusing, and with te removal of #authority
and #ambient, there is now no reason to keep them.

They were a separate kind of string with less operations, that were
interned in the symbol table. That symbol table is used to reduce the
token's size (not repeating strings, referring to them through integers)
and in the Datalog engine for faster unification (it's faster to compare
two integers than two strings).

Now that symbols are gone, the Datalog engine has been modified to use
the symbol table for strings, using the symbols performance boost for
all strings