do not convert variables if we're not build facts or queries
the builder structures are used in multiple ways: to build Datalog
elements to be added to a block, to convert from a token to datalog
engine with symbol table manipulations, and to extract data from the
datalog engine with verifier queries.

Setting variables is only useful in the first case, not the other ones
(we will not set variables on immutables data from the token, nor on
facts obtained from queries), so they should not bear the performance
cost of going through all the elements and replacing the variables.
In those cases, the variables map is not present