add UnverifiedBiscuit type
This represents a token whose signature has not been verified. It is
useful to manipulate a token without full verification: sealing a token
or appending a block can be done without checking the entire signature.
It is also useful in the CLI for analysis.

A verified Biscuit can be obtained from an UnverifiedBiscuit through the
check_signature method