Fix default symbol table
The default symbol table defined in the spec lists `time`, not `current_time`. This is not a big issue for tokens created, attenuated and verified by the same implementation, but it can cause issue for cross implementation use.

The Haskell impl follows the spec, I'm not sure about go and java.