split `error::Token::Sealed` in two errors (#65)
This can happen in two cases:

- trying to append a block to a sealed token (consistent with the error description)
- trying to sealed an already sealed token

When in the second case, the current error description is misleading. Moreover, we might
want treat the two errors differently (ie we should always fail when we can't append a
block, but we might want to ignore the error if we're trying to seal an already sealed
token).

The way `error::Token` is structured makes things a bit awkward (`append` has to catch
and adapt the error returned by `TokenNext.keypair()`).