Param interpolation (#69)
* Add support for datalog parameters

Datalog parameters allow building datalog elements from static strings, combined
with dynamic parameters. Datalog variables were previously co-opted for this, but
this came with several issues:

- variables are not permitted within sets, so it was impossible to make a set element
  dynamic
- forgetting to bind a parameter in a rule body would leave it as a regular variable,
  able to match anything, which is a big footgun
- it is not possible to replace variables with `add_code`

More generally, this conflated two very different things in a single element. Datalog
is already a bit alien for most developers, so grafting a new meaning to datalog variables
was a bit dangerous.

The current design uses a dedicated `validate_parameters` step (same as what was already
done for facts), used when running a builder. This is doubled by a `panic!` in the
builder -> datalog function. I kept the current behaviour, but imo, we should make
`convert` return a `Result` and handle things in a single place.

Co-authored-by: Geoffroy Couprie <contact@geoffroycouprie.com>