3rd-party: add support for scope parameter substitution (runtime & compile-time)
Public keys can now be provided for scopes through parameter
substitution.  The `add_code_with_params` functions now take an
extra parameters hashmap that can hold public keys. Similarly, all
the builder structs carry an extra `scope_parameters` hashmap for
provided parameters.

Macros substitution work a bit differently: term and scope params are
not separated: they can be interleaved, and a dedicated trait projects
them to an enum (term or scope).  This way, the parameter substitution
code knows whether to add them as terms or scope elements.  The initial
plan was to add a blanket Into<Term> => ToAnyParam implementation but
that was not possible because of potential overlapping instances. So
the solution is to complement each Into<Term> impl with a ToAnyParam
impl.

Note that block-level parameter subsitution is not supported yet,
simply because `add_code` functions don't handle it at all either.