remove `Biscuit.create_block()`
It gives the false impression that the block is created from the token,
when it's a brand new block builder, completely untied to the token.
That makes the code surprising, when the block seemingly created from the
token still has to be appended to the token explicitly.

As it is, the function does not pass the fairbairn threshold.