biscuit-quote: add a `rule!` macro
biscuit-quote does not provide macros for individual datalog statements
since the `biscuit`, `block`, `authorizer` and their `_merge` versions
allow for an efficient management of those when manipulating biscuits,
blocks and authorizers.

However, a single `Rule` can be used for querying an authorizer, so we
need a way to easily construct a single one. That's what the `rule!`
macro does. It works the same as the other macros wrt parameter
substitution.