Prevent facts from carrying variables
Facts cannot contain variables. This used to be enforced by the parser,
but this check was relaxed in `27bbc2649ac7f9bad8509aaf6f137efe7d8b0f42`
in order to use variables to perform interpolation. While useful, this
way of using variables caused problems, so in the end it was scrapped
and replaced by a nice interpolation mechanism in #69.

Facts should no longer contain variables at any point, so we can move
the check back to the parser.