update the block rules after symbol translation (#159)
* update the block rules after symbol translation

when loading a token into an authorizer, we must translate the token
data, encoded with the token's symbol table, to authorizer data, encoded
with the authorizer's symbol table, before loading it into the Datalog
world and storing it in the `blocks` field. It was done correctly for
facts and checks, but for rules, the translated version was stored in
the world but not in `blocks`. This had no incidence on authorization,
but meant that snapshots would sometimes store rules with invalid symbols

* clarify internal add_block method