samples: don't use `usize::MAX` for the authorizer block
The rust implementation uses `usize::MAX` as a sentinel value to
represent the authorizer block id. This is specific to the rust impl
and is not part of the spec. The protobuf representation for
snapshot doesn't use it either.

That alone is a good reason to use `null` to denote the authorizer
origin.

Another, more _fun_ reason is that `u32::MAX` cannot be represented
as an IEEE754 double. The Rust JSON implementation, being a sensible
implementation, correctly serializes `u32::MAX` to the corresponding
numeric representation. Other JSON implementations use IEEE754 doubles
to represent JSON numbers (which is not strictly faithful to the JSON
spec (or to the JSON specS, but anyway)). So, when trying to read the
correctly serialized `u32::MAX` value as a `u32` in another language,
which plays fast an loose and uses IEEE754 doubles as the underlying
representation for JSON numbers, deserialization fails (because the
closest IEEE754 number is `u32::MAX + 1`, which is out of bounds).