TTL check: use an empty rule head
Checks bodies use entires rules even though they only care about the
rule body. rules built with parsers use `query()` as the rule body.
Since `query` is part of the default symbol table, it's not costing
any space.

The TTL check uses a manually built check (it cannot use macros,
which are feature-gated), and this check had `expiration($time)` as
the query head, which was both:

- different from how checks are usually represented
- requiring to store `expiration` in the symbol table even though it
  is not used.