feature: add a fact! macro for individual facts
In most cases, we can add facts to a block or an authorizer with the `block_merge!` and `authorizer_merge!` macros, but in some scenarios we still want to manipulate individual facts.

This macro does just that.

Same as the `rule!` macro, it parses a whole block and ensures that it contains exactly one fact, and then emits code to build this fact. The whole block business only happens at compile time, so there is no runtime impact of doing it like this.