authorizer: always display authorizer facts and rules
`impl Display for Authorizer` only used the datalog world for displaying facts. When calling `.to_string()` on an authorizer before `.authorize()`, authorizer facts and rules were not displayed (authorizer facts and rules are added to the datalog world only in `.authorize()`).

After authorization, authorizer facts and rules are both in the world and the block, so deduplication is needed.