wip: datalog foreign function interface prototype
This allows using external functions in datalog.

This makes it easy to provide custom logic without extending the spec for every use-case, at the expense of portability: behaviour is no longer guaranteed to be consistent cross languages, and some languages won’t be able to support it at all (for instance JS as of now).

Todo:
- stricter conversions from datalog
- feature-gating if possible

Open questions:
- enum index for the FFI variants (contiguous or not?)
- how to provide functions (right now, function pointers: prevent mutability and closing over arguments)
- how to provide arguments (right now, datalog::Term, so symbols have to be resolved, and functions returning strings have to register new symbols)