fix version checks; introduce x.y versioning for datalog
While the incrementing integer for version numbers is fine for the protobuf encoding, it has been a headache when communicating with other people.

A `major.minor` scheme seems better for understanding, and helps communicate the scope of changes a bit better.

There was a bug in version checks that relied on `BISCUIT_MAX_VERSION` to check the version for specific features. This does not play well with version upgrades.

So this commit fixes both issues:

- use explicit 3.x version number for datalog features
- use explicit version numbers for feature checks, instead of relying on the max version

It also improves a bit on some error messages that were a bit cryptic, by clearly specifying the version in which specific features were implemented.