crypto: make KeyPair::new() default on using ed25519
This has been the default behaviour since biscuit 2. Taking the algorithm as a parameter is a breaking change which is IMO not necessary. This will make the update a little bit less painful for consumers.

Also, I don’t know the consensus on having `Default::default()` not return the same value every time, but I find it a bit misleading.