Add pem/der loaders to PublicKey (#265)
* Add pem/der loaders to PublicKey
Expose pem feature in biscuit-parser crate as well

* missing deserialization methods

* serialization methods

* fix

* more tests

* allow guessing the algorithm from pem/der keypairs / public keys / private keys

The pem/der payloads contain information about the key algorithm. Unfortunately we can’t easily extract it ahead of time, but we can try for each supported algorithms. Key parsing is typically a one-time operation, so it’s ok to do it this way.

* rename functions

---------

Co-authored-by: baranyildirim@gmail.com <baran.yildirim@outlook.com>
Co-authored-by: Clement Delafargue <clement.delafargue@outscale.com>